Add files via upload
This is code taken from examples in a workbook. It is NOT my work and I am not affiliated with the authors, but I highly recommend using this book if you want an easy to follow intro to Machine Learning in Python.

The ebook is available for free here: https://play.google.com/store/books/details/Lisa_Tagliaferri_Python_Machine_Learning_Projects?id=YU2eDwAAQBAJ


The program:
GaussianNB_Tumor_Prediction.py takes the examples from the book and puts it into one program. The end result is a classifier script that loads data, organizes it, displays it, then makes a prediction of whether a tumour is malignant or benign. It then evaluates the result at around 94% accuracy. 

Modules/requirements: SciKit (pip install sklearn)
Preferred environment: Jupyter Notepad however this version will run in IDLE provided the dependencies are met.

There are some other good projects in the book too.